FixGtkdocFilter is not a filter anymore
FixGtkdocFilter can't be an HTML::Pipeline filter anymore; this commit
turns it into a preprocessor special to the Mallard scraper. This is
because the HTML::Pipeline module loses whitespace information when
parsing the HTML whereas we need to treat some sections like CDATA (e.g.
code between |[...]| needs newlines intact.

Also, the GtkDoc markup we are trying to fix up can span multiple HTML
elements. An example is where a program listing between |[...]| contains
a blank line; that will get converted into the end of a paragraph.