Convert |[...]| into <pre><code>
This converts GtkDoc's special markup |[...]| for a program listing into
HTML <pre><code> elements. It somewhat lamely adds a 'language' attribute
to the <code> element if there was a comment at the start of the code
block stating what language the code block was in.

It also removes stray <p> elements and other HTML elements from the code
because Mallard replaces a double-newline with a paragraph break. (It
doesn't need to worry about XML elements in the code block itself,
because those will all be escaped with &lt; and &gt;.