Add GIR docs icons
This adds some icons for GTK and Gnome stack documentation.

Since new sprites get added to master quite often, we will undoubtedly
have to deal with sprite merge conflicts if we incorporate changes from
master. Instead, put all the GIR docs icons into a separate sprite, and
use those for all the current documentation sets that we have.

As well, some of the icons should be reversed in the dark theme.