Initial attempt at generating HTML directly
This is an attempt to avoid using g-ir-doc-tool to go from GIR XML
documentation -> Mallard XML -> HTML without metadata -> HTML with
metadata. Instead, we generate some HTML with all the metadata we need,
directly from the GIR XML.

This has the added benefit of being able to render the Markdown
more accurately instead of "fixing" it with regexes.

This is a minimum implementation that doesn't work very well yet. It has
many regressions relative to the previous approach.