Add CSS overrides for GIR-generated pages
DevDocs has a CSS rule somewhere that code inside <code> elements should
have its whitespace preserved as if it were inside a <pre> element. In
the GIR-generated pages we only want this if the <code> element is
actually inside a <pre> element.

We do this with a CSS override file, which has the added benefit of
applying the default CSS to the documentation, making it look quite a lot
better.