Put class name in sidebar for additional entries
Even though this makes some entries take up more space than there is in
the sidebar, it's still better if you want to search for
GdkPixbuf.Pixbuf.new(), for example. Without it, the entry you're
searching for is just "new()", one of many other entries. You can't type
"pixbufnew" to reach it either. With the class name prepended, you can.

Separate signals from the class name with '::', for lack of something
better; everything else gets a '.'. Signals are treated differently
because you can't access them as members in Javascript.